Alter color scales for Bar, Line, Pie, Radar, Sankey and Treemap (#159)
Most of them are now using `d3.scale.category10()`, while the Bar Chart and
Treeemap use a constant color (blueish, the first color of the category scale).
The remaining charts are the Map and Bubble Tree. Both weren't modified now
because they require a different treatment, as the Map doesn't use a categorical
scale, and the Bubble Tree doesn't use D3.