[#1274] Fix bug where we called encodeURIComponent() twice (#164)
We should only call it once, otherwise a string like `date:2017` becomes
`date%3A2017` in the first call, and `date%253A2017` in the second call. In
other words, `encodeURIComponent()` encodes the `%` in `date%3A2017`.

The reasoning here is that `serializeCut()` should only serialize the cuts, not
bothering with encoding URIs or not. We should only encode just before creating
the HTTP requests. Hopefully, this will avoid we unknowingly encoding the values
multiple times.

Fixes openspending/openspending#1274