[#1318] Fix negative number formatting
There's still a remaining issue for numbers in the extremes of our suffixes
(thousand, million, etc.). For example, 999999 should either become '999.99
Thousand' or '1.0 Million'. Instead, our current code turns that into '1000.0
Million'.

This is a tricky bug, caused by the fact that `toFixed()` also rounds numbers.
I don't have time to fix it now, and it shouldn't affect many people. I added a
skipped test that triggers the bug for when we're able to fix it.

Fixes openspending/openspending#1318