Fix init to prevent broken redirect
The configuration has to be loaded before being able to access its
properties. This commit ensures that if there are no datapackages, the
redirect to os-explorer will work. It will also prevent excecution of
the then chain (causing errors because there is no datapackage) by
nesting it in an else clause.