Default paramater controls for views
Adds controls for dataset owner to save current view as default. Add
user controls to reset the entire view, or reset the view to the default
parameters.

- Adds a `param-controls` directive to manage the UI and controller code
for the default params.
- View params are saved via the data-package-api to the package model in
the ElasticSearch instance. This is mediated by os-conductor by the new
`/package/update_params` endpoint.
- `confirmation-popover` directive added to UI for the 'Save View As
Default' action, which is only shown to the dataset owner.