Send param as string value.
Because Elasticsearch update merges object values with existing values,
we need a way to replace the params entirely. We don't have access to
Elasticsearch scripting, so the easiest way is to dump the params object
to a string, and send it as a value of a 'params' key. The stringified
json object is parsed back to an object before use.